+++ CHANGELOG.md
@@ -10,6 +10,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
 - Add support for deep match on filter expression predicate, exemple selector: `$.addresses[?(@.address.state=='Bahia')]`
 
+### Fixed
+- Fix DescendantOperator weren't apply the filter predicate properly when it's operate on first-class citizen data type.
+Example: Queries like `Warpath.query([[1, 2], :item, 9, [9, 8]], "$..[?( is_list(@) )]")`, always returned empty list, now it will produce, `[[1, 2], [9, 8]]`
+
 ## [0.3.0] - 2020-06-09
 This release is a complete new implementation strategy, it relay on elixir protocol to promote extensibility and simplify maintainability.